♻️ refactor: 공통 코드 수정으로 인한 리팩토링
