+++ src/main/java/com/gogym/common/response/ApplicationResponse.java
@@ -1,6 +1,7 @@
 package com.gogym.common.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import com.gogym.exception.ErrorCode;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -14,19 +15,16 @@ public class ApplicationResponse<T> {
   private final String responseCode;
   private final String responseMessage;
 
-  public static <T> ApplicationResponse<T> ok(T data, SuccessCode code) {
-    return new ApplicationResponse<>(data, code.getCode(), code.getMessage());
-  }
 
-  public static ApplicationResponse<Void> noData(SuccessCode code) {
-    return new ApplicationResponse<>(null, code.getCode(), code.getMessage());
-  }
+  public static <T> ApplicationResponse<T> ok(T data, String code, String message) {
+    return new ApplicationResponse<>(data, code, message);
+    }
 
-  public static ApplicationResponse<Void> error(ErrorCode code) {
-    return new ApplicationResponse<>(null, code.getCode(), code.getMessage());
-  }
+  public static ApplicationResponse<Void> noData(String code, String message) {
+    return new ApplicationResponse<>(null, code, message);
+    }
 
   public static ApplicationResponse<Void> error(String code, String message) {
     return new ApplicationResponse<>(null, code, message);
-  }
+    }
 }
