--- src/main/java/com/gogym/common/response/ResponseCode.java
@@ -1,12 +1,0 @@
-package com.gogym.common.response;
-
-import org.springframework.http.HttpStatus;
-
-public interface ResponseCode {
-
-  HttpStatus getHttpStatus();
-
-  String getCode();
-
-  String getMessage();
-}
