+++ src/main/java/com/gogym/exception/CustomException.java
@@ -1,6 +1,5 @@
 package com.gogym.exception;
 
-import com.gogym.common.response.ErrorCode;
 import lombok.Getter;
 
 @Getter
