+++ src/main/java/com/gogym/member/controller/AuthController.java
@@ -1,14 +1,12 @@
 package com.gogym.member.controller;
 
 import com.gogym.common.response.ApplicationResponse;
-import com.gogym.common.response.SuccessCode;
 import com.gogym.member.dto.SignUpRequest;
 import com.gogym.member.dto.SignInRequest;
 import com.gogym.member.dto.ResetPasswordRequest;
 import com.gogym.member.dto.AuthResponse;
 import com.gogym.member.service.AuthService;
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotBlank;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
@@ -22,16 +20,20 @@ import java.util.List;
 public class AuthController {
 
   private static final String ROLE_USER = "ROLE_USER";
+  private static final String SUCCESS_CODE = "200";
+  private static final String SUCCESS_MESSAGE = "요청이 성공적으로 처리되었습니다.";
+  private static final String DATA_AVAILABLE_MESSAGE = "요청하신 데이터는 사용 가능합니다.";
+  private static final String EMAIL_VERIFIED_MESSAGE = "이메일 인증이 완료되었습니다.";
 
   private final AuthService authService;
 
   // 회원가입
   @PostMapping("/sign-up")
   public ResponseEntity<ApplicationResponse<AuthResponse>> signUp(
-    @RequestBody @Valid SignUpRequest request // 요청 데이터 검증 및 전달
+    @RequestBody @Valid SignUpRequest request
   ) {
-    AuthResponse response = authService.signUp(request); // 회원가입 처리
-    return ResponseEntity.ok(ApplicationResponse.ok(response, SuccessCode.AUTH_SUCCESS)); // 성공 응답 반환
+    AuthResponse response = authService.signUp(request);
+    return ResponseEntity.ok(ApplicationResponse.ok(response, SUCCESS_CODE, SUCCESS_MESSAGE));
   }
 
   // 로그인
@@ -39,16 +41,12 @@ public class AuthController {
   public ResponseEntity<ApplicationResponse<AuthResponse>> login(
     @RequestBody @Valid SignInRequest request
   ) {
-    // 로그인 처리 및 사용자 정보 생성
     AuthResponse response = authService.login(request);
-
-    // JWT 토큰 생성
     String token = authService.generateToken(request.getEmail(), List.of(ROLE_USER));
 
-    // 응답
     return ResponseEntity.ok()
       .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
-      .body(ApplicationResponse.ok(response, SuccessCode.AUTH_SUCCESS));
+      .body(ApplicationResponse.ok(response, SUCCESS_CODE, SUCCESS_MESSAGE));
   }
 
   // 로그아웃
@@ -57,46 +55,37 @@ public class AuthController {
     @RequestHeader("Authorization") String token
   ) {
     authService.logout(token.replace("Bearer ", ""));
-    return ResponseEntity.ok(ApplicationResponse.noData(SuccessCode.AUTH_SUCCESS));
+    return ResponseEntity.ok(ApplicationResponse.noData(SUCCESS_CODE, SUCCESS_MESSAGE));
   }
 
   // 비밀번호 재설정
   @PostMapping("/reset-password")
   public ResponseEntity<ApplicationResponse<Void>> resetPassword(
-    @RequestBody @Valid ResetPasswordRequest request // 요청 데이터 검증 및 전달
+    @RequestBody @Valid ResetPasswordRequest request
   ) {
-    authService.resetPassword(request); // 비밀번호 재설정 처리
-    return ResponseEntity.ok(ApplicationResponse.noData(SuccessCode.AUTH_SUCCESS)); // 성공 응답 반환
+    authService.resetPassword(request);
+    return ResponseEntity.ok(ApplicationResponse.noData(SUCCESS_CODE, SUCCESS_MESSAGE));
   }
 
   // 이메일 중복 확인
   @GetMapping("/check-email")
   public ApplicationResponse<Void> checkEmail(@RequestParam("email") String email) {
-    // 이메일 중복 확인 서비스 호출
     authService.checkEmail(email);
-
-    // 성공 응답 반환
-    return ApplicationResponse.noData(SuccessCode.DATA_AVAILABLE);
+    return ApplicationResponse.noData(SUCCESS_CODE, DATA_AVAILABLE_MESSAGE);
   }
 
   // 닉네임 중복 확인
   @GetMapping("/check-nickname")
   public ApplicationResponse<Void> checkNickname(@RequestParam("nickname") String nickname) {
-    // 닉네임 중복 확인 서비스 호출
     authService.checkNickname(nickname);
-
-    // 성공 응답 반환
-    return ApplicationResponse.noData(SuccessCode.DATA_AVAILABLE);
+    return ApplicationResponse.noData(SUCCESS_CODE, DATA_AVAILABLE_MESSAGE);
   }
 
   // 이메일 인증 확인
   @GetMapping("/verify-email")
   public ApplicationResponse<Void> verifyEmail(@RequestParam(name = "token") String token) {
-    // 이메일 인증 토큰 검증 서비스 호출
     authService.verifyEmailToken(token);
-
-    // 성공 응답 반환
-    return ApplicationResponse.noData(SuccessCode.EMAIL_VERIFIED);
+    return ApplicationResponse.noData(SUCCESS_CODE, EMAIL_VERIFIED_MESSAGE);
   }
 
   // 이메일 인증 요청
@@ -104,10 +93,11 @@ public class AuthController {
   public ResponseEntity<ApplicationResponse<Void>> sendVerificationEmail(
       @RequestParam(name = "email") String email
   ) {
-      authService.sendVerificationEmail(email);
-      return ResponseEntity.ok(ApplicationResponse.noData(SuccessCode.AUTH_SUCCESS));
+    authService.sendVerificationEmail(email);
+    return ResponseEntity.ok(ApplicationResponse.noData(SUCCESS_CODE, SUCCESS_MESSAGE));
   }
 }
+
